+++ src/ngx-dnd.module.ts
@@ -1,7 +1,5 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { BrowserModule } from '@angular/platform-browser';
-import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
 import { DraggableDirective, DroppableDirective, DragHandleDirective } from './directives';
 import { ContainerComponent, ItemComponent } from './components';
@@ -12,8 +10,6 @@ const directives = [DraggableDirective, DroppableDirective, DragHandleDirective]
 
 @NgModule({
   imports: [
-    BrowserModule,
-    BrowserAnimationsModule,
     CommonModule
   ],
   declarations: [